windowing/egl/wayland: move to windowing/wayland/
In https://github.com/xbmc/xbmc/pull/8980, FernetMenta postulated:

 "If anybody wants to re-submit wayland, it has to comply to the
 architecture.  That is no components under the folder windowing/egl,
 no boost."

This patch is trying to comply with this rule.
