+++ xbmc/windowing/tests/wayland/TestXBMCWaylandInputAcceptance.cpp
@@ -29,19 +29,19 @@
 #include <wayland-client-protocol.h>
 #include <wayland-version.h>
 #include "xbmc_wayland_test_client_protocol.h"
-#include "windowing/DllWaylandClient.h"
-#include "windowing/DllWaylandEgl.h"
-#include "windowing/DllXKBCommon.h"
+#include "windowing/wayland/DllWaylandClient.h"
+#include "windowing/wayland/DllWaylandEgl.h"
+#include "windowing/wayland/DllXKBCommon.h"
 
-#include "windowing/egl/wayland/Callback.h"
-#include "windowing/egl/wayland/Compositor.h"
-#include "windowing/egl/wayland/Display.h"
-#include "windowing/egl/wayland/OpenGLSurface.h"
-#include "windowing/egl/wayland/Registry.h"
-#include "windowing/egl/wayland/Surface.h"
-#include "windowing/egl/wayland/Shell.h"
-#include "windowing/egl/wayland/ShellSurface.h"
-#include "windowing/egl/EGLNativeTypeWayland.h"
+#include "windowing/wayland/Callback.h"
+#include "windowing/wayland/Compositor.h"
+#include "windowing/wayland/Display.h"
+#include "windowing/wayland/OpenGLSurface.h"
+#include "windowing/wayland/Registry.h"
+#include "windowing/wayland/Surface.h"
+#include "windowing/wayland/Shell.h"
+#include "windowing/wayland/ShellSurface.h"
+#include "windowing/wayland/EGLNativeTypeWayland.h"
 #include "windowing/wayland/EventLoop.h"
 #include "windowing/wayland/EventQueueStrategy.h"
 #include "windowing/wayland/TimeoutManager.h"
