+++ xbmc/windowing/wayland/Compositor.h
