+++ xbmc/windowing/wayland/Display.cpp
@@ -25,8 +25,8 @@
 
 #include <wayland-client.h>
 
-#include "windowing/DllWaylandClient.h"
-#include "windowing/WaylandProtocol.h"
+#include "DllWaylandClient.h"
+#include "WaylandProtocol.h"
 #include "Display.h"
 
 namespace xw = xbmc::wayland;
