+++ xbmc/windowing/wayland/Display.h
