+++ xbmc/windowing/wayland/DllWaylandClient.h
