+++ xbmc/windowing/wayland/DllWaylandEgl.h
