+++ xbmc/windowing/wayland/Output.h
