+++ xbmc/windowing/wayland/Registry.cpp
@@ -19,8 +19,8 @@
  */
 #include <wayland-client.h>
 
-#include "windowing/DllWaylandClient.h"
-#include "windowing/WaylandProtocol.h"
+#include "DllWaylandClient.h"
+#include "WaylandProtocol.h"
 #include "Registry.h"
 
 namespace xw = xbmc::wayland;
