+++ xbmc/windowing/wayland/Shell.h
