+++ xbmc/windowing/wayland/Wayland12EventQueueStrategy.cpp
@@ -23,7 +23,7 @@
 #define _GNU_SOURCE
 #endif
 
-#include "windowing/DllWaylandClient.h"
+#include "DllWaylandClient.h"
 #include "utils/log.h"
 
 #include "Wayland12EventQueueStrategy.h"
