+++ xbmc/windowing/wayland/WaylandLibraries.cpp
