+++ xbmc/windowing/wayland/WaylandProtocol.h
