+++ xbmc/windowing/wayland/WinEventsWayland.h
