+++ xbmc/windowing/wayland/XBMCConnection.h
